requirements.txt
aiogram
sqlalchemy
asyncpg
python-dotenv
openai
ijson
